requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiogram==3.18.0
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
pywin32==310
